Remove redundant “rich” component
Addresses feedback in https://github.com/WordPress/gutenberg/pull/54718#discussion_r1404261330
